+++ src/UI/Components/BasicInfo/BasicInfo/BasicInfo.js
@@ -85,7 +85,7 @@ define(function(require)
 		this.ui.find('.buttons button').mousedown(function(){
 			switch (this.className) {
 				case 'item':
-					Inventory.ui.toggle();
+					Inventory.getUI().toggle();
 					break;
 
 				case 'info':
