+++ src/UI/Components/BasicInfo/BasicInfoV5/BasicInfoV5.js
@@ -92,7 +92,7 @@ define(function(require)
 		this.ui.find('.buttons div').mousedown(function(){
 			switch (this.id) {
 				case 'item':
-					Inventory.ui.toggle();
+					Inventory.getUI().toggle();
 					break;
 
 				case 'info':
