--- src/UI/Components/Inventory/Inventory.css
@@ -1,41 +1,0 @@
-#Inventory { position:absolute; top:100px; left:100px; font-size:12px; }
-#Inventory table { border-spacing:0px; display:inline-block; }
-
-#Inventory .titlebar { width:100%; height:17px; background-color:white; background-repeat:repeat-x; border-radius:3px 3px 0px 0px; }
-#Inventory .titlebar .base { width:11px; height:11px; border:none; background-color:transparent; background-repeat:no-repeat; vertical-align:middle; }
-#Inventory .titlebar .text { text-shadow:1px 1px white; vertical-align:-2px; white-space:nowrap;
- /* chrome bug */ display: inline-block; width: 32px; height:13px; }
-
-#Inventory .titlebar .left { margin-left:3px; float:left; }
-#Inventory .titlebar .right { float:right; margin-right:3px;}
-#Inventory .titlebar .clear { clear:both; }
-
-#Inventory .tabs { border-left:1px solid #ccc; width:50px; background-repeat:no-repeat; background-color:white; vertical-align:top; background-position: -1px 0px; }
-#Inventory .tabs button { width:20px; height:27px; border:none; background-color:transparent; display:block; }
-
-#Inventory .container { padding-left: 14px; border-right:1px solid #ccc; background:white; }
-#Inventory .ff_bugfix { position:relative; width: 100%; height: 100%; }
-#Inventory .hide { height:100%; width:16px; position:absolute; top:0px; right:0px; background-color:white;}
-#Inventory .content { overflow:auto; width:100%; height:100%; min-height:65px; background-color:transparent; background-repeat:repeat;  background-attachment: local;}
-
-#Inventory .content .item { display:block; width:24px; height:24px; margin:4px 4px 4px 4px; position:relative; float:left; }
-#Inventory .content .item .icon { width:24px; height:24px; border:none; background-color:transparent; background-repeat:no-repeat; }
-#Inventory .overlay {
-    position: absolute;
-    display: none;
-    white-space: nowrap;
-    z-index: 900;
-    height: 15px;
-    line-height: 15px;
-    border-radius: 3px;
-    padding: 4px;
-    background: rgba(0,0,0,0.7);
-    color: white;
-    text-shadow: 1px 1px black;
-}
-#Inventory .overlay.grey { color:#aaa; }
-#Inventory .content .item .amount { position:relative; font-size:11px; bottom:9px; right:0px; text-align:right; text-shadow:-1px -1px white; }
-
-#Inventory .footer { width:100%; height:27px; background-repeat:repeat-x; background-color:transparent; position:relative; border-right:1px solid #ccc; }
-#Inventory .footer .cnt { position:absolute; left:10px; bottom:6px; font-size:12px; }
-#Inventory .footer button { position:absolute; right:0px; bottom:1px; width:13px; height:13px; border:none; background-repeat:no-repeat; background-color:transparent; }
