+++ src/UI/Components/Inventory/InventoryV0/InventoryV0.css
@@ -1,0 +1,41 @@
+#InventoryV0 { position:absolute; top:100px; left:100px; font-size:12px; }
+#InventoryV0 table { border-spacing:0px; display:inline-block; }
+
+#InventoryV0 .titlebar { width:100%; height:17px; background-color:white; background-repeat:repeat-x; border-radius:3px 3px 0px 0px; }
+#InventoryV0 .titlebar .base { width:11px; height:11px; border:none; background-color:transparent; background-repeat:no-repeat; vertical-align:middle; }
+#InventoryV0 .titlebar .text { text-shadow:1px 1px white; vertical-align:-2px; white-space:nowrap;
+ /* chrome bug */ display: inline-block; width: 32px; height:13px; }
+
+#InventoryV0 .titlebar .left { margin-left:3px; float:left; }
+#InventoryV0 .titlebar .right { float:right; margin-right:3px;}
+#InventoryV0 .titlebar .clear { clear:both; }
+
+#InventoryV0 .tabs { border-left:1px solid #ccc; width:50px; background-repeat:no-repeat; background-color:white; vertical-align:top; background-position: -1px 0px; }
+#InventoryV0 .tabs button { width:20px; height:27px; border:none; background-color:transparent; display:block; }
+
+#InventoryV0 .container { padding-left: 14px; border-right:1px solid #ccc; background:white; }
+#InventoryV0 .ff_bugfix { position:relative; width: 100%; height: 100%; }
+#InventoryV0 .hide { height:100%; width:16px; position:absolute; top:0px; right:0px; background-color:white;}
+#InventoryV0 .content { overflow:auto; width:100%; height:100%; min-height:65px; background-color:transparent; background-repeat:repeat;  background-attachment: local;}
+
+#InventoryV0 .content .item { display:block; width:24px; height:24px; margin:4px 4px 4px 4px; position:relative; float:left; }
+#InventoryV0 .content .item .icon { width:24px; height:24px; border:none; background-color:transparent; background-repeat:no-repeat; }
+#InventoryV0 .overlay {
+    position: absolute;
+    display: none;
+    white-space: nowrap;
+    z-index: 900;
+    height: 15px;
+    line-height: 15px;
+    border-radius: 3px;
+    padding: 4px;
+    background: rgba(0,0,0,0.7);
+    color: white;
+    text-shadow: 1px 1px black;
+}
+#InventoryV0 .overlay.grey { color:#aaa; }
+#InventoryV0 .content .item .amount { position:relative; font-size:11px; bottom:9px; right:0px; text-align:right; text-shadow:-1px -1px white; }
+
+#InventoryV0 .footer { width:100%; height:27px; background-repeat:repeat-x; background-color:transparent; position:relative; border-right:1px solid #ccc; }
+#InventoryV0 .footer .cnt { position:absolute; left:10px; bottom:6px; font-size:12px; }
+#InventoryV0 .footer button { position:absolute; right:0px; bottom:1px; width:13px; height:13px; border:none; background-repeat:no-repeat; background-color:transparent; }
