+++ src/UI/Components/Inventory/InventoryV0/InventoryV0.html
@@ -1,4 +1,4 @@
-<div id="Inventory">
+<div id="InventoryV0">
 	<div class="titlebar" data-background="basic_interface/titlebar_mid.bmp">
 		<div class="left">
 			<button class="base" data-background="basic_interface/sys_base_off.bmp" data-hover="basic_interface/sys_base_on.bmp"></button>
