+++ src/UI/Components/Mail/Mail.js
@@ -634,7 +634,7 @@
 
 				if(data.from == 'Inventory')
 				{
-					Inventory.removeItem(
+					Inventory.getUI().removeItem(
 						item.index,
 						parseInt(count, 10 )
 					);		
@@ -655,7 +655,7 @@
 
 		
 		if(data.from == 'Inventory'){
-			Inventory.removeItem( item.index, 1 );						
+			Inventory.getUI().removeItem( item.index, 1 );						
 		}
 		Mail.parseMailWinopen(1); // remove item
